+++ truffle/src/com.oracle.truffle.sl.test/src/com/oracle/truffle/sl/test/SLInteropPrimitiveTest.java
@@ -40,6 +40,8 @@
  */
 package com.oracle.truffle.sl.test;
 
+import java.math.BigInteger;
+
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.Source;
 import org.graalvm.polyglot.Value;
@@ -75,5 +77,18 @@ public class SLInteropPrimitiveTest {
         final Value fnc = context.eval(src);
         Assert.assertTrue(fnc.canExecute());
         fnc.execute('a', 'b');
+    }
+
+    @Test
+    public void testNumbers() {
+        final Source src = Source.newBuilder("sl", "function addNum(a,b) {return a + b;} function main() {return addNum;}", "addNum.sl").buildLiteral();
+        final Value fnc = context.eval(src);
+        Assert.assertTrue(fnc.canExecute());
+        Assert.assertEquals(42, fnc.execute(20, 22).asInt());
+        Assert.assertEquals(new BigInteger("18446744073709551614"), fnc.execute(Long.MAX_VALUE, Long.MAX_VALUE).asBigInteger());
+        Assert.assertEquals(new BigInteger("18446744073709551616"), fnc.execute(BigInteger.valueOf(Long.MAX_VALUE).add(BigInteger.TWO), Long.MAX_VALUE).asBigInteger());
+        Assert.assertEquals(new BigInteger("18446744073709551616"), fnc.execute(BigInteger.valueOf(Long.MAX_VALUE), BigInteger.valueOf(Long.MAX_VALUE).add(BigInteger.TWO)).asBigInteger());
+        Assert.assertEquals(new BigInteger("18446744073709551618"),
+                        fnc.execute(BigInteger.valueOf(Long.MAX_VALUE).add(BigInteger.TWO), BigInteger.valueOf(Long.MAX_VALUE).add(BigInteger.TWO)).asBigInteger());
     }
 }