+++ truffle/src/com.oracle.truffle.sl/src/com/oracle/truffle/sl/SLLanguage.java
@@ -106,7 +106,7 @@ import com.oracle.truffle.sl.nodes.local.SLWriteLocalVariableNode;
 import com.oracle.truffle.sl.parser.SLNodeFactory;
 import com.oracle.truffle.sl.parser.SimpleLanguageLexer;
 import com.oracle.truffle.sl.parser.SimpleLanguageParser;
-import com.oracle.truffle.sl.runtime.SLBigNumber;
+import com.oracle.truffle.sl.runtime.SLBigInteger;
 import com.oracle.truffle.sl.runtime.SLContext;
 import com.oracle.truffle.sl.runtime.SLFunction;
 import com.oracle.truffle.sl.runtime.SLFunctionRegistry;
@@ -130,7 +130,7 @@ import com.oracle.truffle.sl.runtime.SLStrings;
  * <b>Types:</b>
  * <ul>
  * <li>Number: arbitrary precision integer numbers. The implementation uses the Java primitive type
- * {@code long} to represent numbers that fit into the 64 bit range, and {@link SLBigNumber} for
+ * {@code long} to represent numbers that fit into the 64 bit range, and {@link SLBigInteger} for
  * numbers that exceed the range. Using a primitive type such as {@code long} is crucial for
  * performance.
  * <li>Boolean: implemented as the Java primitive type {@code boolean}.